JUnit port for IsInstanceOfTest
